+++ src/index.ts
@@ -1,7 +1,5 @@
-export { AudioFrameBufferWriter } from './audio-frame-buffer-writer'
-export { AudioFrameBufferReader } from './audio-frame-buffer-reader'
-export {
-  type AudioFrameBufferContext,
-  createAudioFrameBufferContext } from './audio-frame-buffer-context'
-export type { AudioFrameBufferParams } from './audio-frame-buffer-params'
-export type { AudioFrameSegment } from './audio-frame-segment'
+export { FrameBufferWriter } from './frame-buffer-writer'
+export { FrameBufferReader } from './frame-buffer-reader'
+export { type FrameBufferContext, createFrameBufferContext } from './frame-buffer-context'
+export type { FrameBufferParams } from './frame-buffer-params'
+export type { FrameBufferSegment } from './frame-buffer-segment'